+++ tests/ServiceLocator/Resolver/Reflection/ReflectionA.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace ExtendsSoftware\ExaPHP\ServiceLocator\Resolver\Reflection;
 
 use ExtendsSoftware\ExaPHP\ServiceLocator\ServiceLocatorInterface;
+use ExtendsSoftware\ExaPHP\Utility\Container\ContainerInterface;
 
 class ReflectionA
 {
@@ -12,8 +13,12 @@ class ReflectionA
      *
      * @param ReflectionB             $b
      * @param ServiceLocatorInterface $serviceLocator
+     * @param ContainerInterface      $container
      */
-    public function __construct(ReflectionB $b, ServiceLocatorInterface $serviceLocator)
-    {
+    public function __construct(
+        ReflectionB             $b,
+        ServiceLocatorInterface $serviceLocator,
+        ContainerInterface      $container
+    ) {
     }
 }